requirements.txt
click==8.1.3
Flask==2.2.3
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
Werkzeug==2.2.3
